Refactored Actions structure
