+++ src/actions/layout/loading/unsetLoading.js
@@ -1,4 +1,4 @@
-import { LAYOUT_UNSET_LOADING } from "./actions";
+import { LAYOUT_UNSET_LOADING } from "../actions";
 
 /**
  * Sets the global layout state as not loading
