+++ src/actions/switcheo/network/toggleNetwork.js
@@ -1,11 +1,11 @@
-import { SWITCHEO_TOGGLE_NETWORK } from "./actions";
+import { SWITCHEO_TOGGLE_NETWORK } from "../actions";
 
-import { fetchContracts } from './fetchContracts'
-import { fetchTokens } from './fetchTokens'
+import { fetchContracts } from '../contracts/fetchContracts'
+import { fetchTokens } from '../tokens/fetchTokens'
 
-import { setAddress } from "../address/setAddress";
-import { raiseError } from "../layout/raiseError";
-import { mainNet} from "../../reducers/switcheo/networkReducer";
+import { setAddress } from "../../address/address/setAddress";
+import { raiseError } from "../../layout/errors/raiseError";
+import { mainNet} from "../../../reducers/switcheo/networkReducer";
 
 /**
  * toggles the current network between test-api and prod-api
